refactor(overlay): split OverlayWidget into hooks + presentational components
Pure structural split of app/src/components/OverlayWidget.tsx along the
seams PR1 (geometry contract) and PR2 (expansion controller) established.
Zero behavior/visual change, no Rust changes.

- deriveVisual.ts: pure top-bar indicator + flash-priority derivation,
  locked by an exhaustive status x flag matrix test.
- previewPresentation.ts: relocated (unchanged) below-notch preview logic.
- useOverlayRuntime / useOverlaySettingsMirror / useRecordingControls /
  useWaveform: extracted Tauri listeners, settings mirror + quick controls,
  click/double-click handling, and the rAF waveform animation.
- useRecordingControls now reads the microphone override via loadSettings()
  instead of a raw localStorage.getItem(STORAGE_KEY) parse (issue #280
  decision 5).
- OverlayPill / OverlayDropdown: presentational split of the top bar and
  quick-settings dropdown; icons colocated with their consuming component.
- OverlayWidget.tsx is now a ~150-line composition shell.

Docs: full rewrite of docs/features/overlay.md to match the new
architecture (also fixes the stale "entire overlay surface is a drag
region" contradiction and de-duplicates repeated geometry/motion numbers).
Also fixes the PR1-carried stale references (get_notch_info,
notch-info-changed) in docs/reference/commands.md, docs/reference/events.md,
and docs/ARCHITECTURE.md, and updates the CLAUDE.md File Map.
